requirements.txt
Django==2.1.4
django-appconf==1.0.2
django-extensions==2.1.4
django-imagekit==4.0.2
django-js-asset==1.1.0
django-mptt==0.9.1
django-seed==0.1.9
Faker==1.0.1
mysqlclient==1.3.14
pilkit==2.0
Pillow==5.3.0
python-dateutil==2.7.5
pytz==2018.7
six==1.12.0
text-unidecode==1.2
